+++ examples/async-graphql/src/events/handler.rs
@@ -7,25 +7,29 @@ use nakago_axum::auth::Subject;
 use nakago_derive::Provider;
 use nakago_ws::connections::Connections;
 
-use crate::domains::users::{self, model::User};
+use crate::domains::users;
 
-use super::messages::{IncomingMessage, OutgoingMessage};
+use super::{
+    messages::{IncomingMessage, OutgoingMessage},
+    session::Session,
+};
 
 /// Message Handler
 #[derive(Clone)]
 pub struct Handler {
-    connections: Arc<Connections<User>>,
+    connections: Arc<Connections<Session>>,
     users: Arc<Box<dyn users::Service>>,
 }
 
 #[async_trait]
-impl nakago_ws::Handler<User> for Handler {
-    async fn get_user(&self, sub: Subject) -> Option<User> {
+impl nakago_ws::Handler<Session> for Handler {
+    async fn get_session(&self, sub: Subject) -> Option<Session> {
         if let Subject(Some(ref username)) = sub {
             self.users
                 .get_by_username(username, &true)
                 .await
-                .unwrap_or(None)
+                .map(Session::new)
+                .ok()
         } else {
             None
         }
@@ -58,12 +62,12 @@ pub struct Provide {}
 
 #[Provider]
 #[async_trait]
-impl Provider<Box<dyn nakago_ws::Handler<User>>> for Provide {
+impl Provider<Box<dyn nakago_ws::Handler<Session>>> for Provide {
     async fn provide(
         self: Arc<Self>,
         i: nakago::Inject,
-    ) -> provider::Result<Arc<Box<dyn nakago_ws::Handler<User>>>> {
-        let connections = i.get::<Connections<User>>().await?;
+    ) -> provider::Result<Arc<Box<dyn nakago_ws::Handler<Session>>>> {
+        let connections = i.get::<Connections<Session>>().await?;
         let users = i.get::<Box<dyn users::Service>>().await?;
 
         Ok(Arc::new(Box::new(Handler { connections, users })))